apps/worker/requirements.txt
# Stream2Short Worker Dependencies

# Redis client
redis>=5.0.0

# Supabase client
supabase>=2.0.0

# HTTP client
httpx>=0.25.0

# Whisper for transcription (local fallback)
faster-whisper>=0.10.0

# Groq API for fast cloud transcription
groq>=0.4.0

# Google Gemini for vision AI (webcam detection)
google-genai>=1.0.0

# Google Drive API
google-api-python-client>=2.111.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0

# Video downloading (Twitch clips)
yt-dlp>=2024.1.0

# Face/webcam detection (contrib version includes trackers: CSRT, KCF, MOSSE)
opencv-contrib-python-headless>=4.9.0
numpy>=1.26.0

# YOLOv8 webcam detector (webcam_locator package — fine-tuned on Twitch clips)
ultralytics>=8.0.0

# Speaker diarization (optional - requires HF_TOKEN and model access)
# Using CPU-only PyTorch to avoid large NVIDIA CUDA downloads
# Version compatibility: pyannote.audio 3.x + huggingface-hub <1.0 (uses use_auth_token)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchaudio==2.5.1
huggingface-hub<1.0.0
pyannote.audio>=3.1.0,<4.0.0

# Audio analysis for gender detection (pitch analysis)
librosa>=0.10.0
soundfile>=0.12.0

